Added login session management

+++ frontend/src/api/apiServices.js
@@ -1,6 +1,8 @@
 import axios from 'axios'
 
 const server = 'api/v1/'
+import state from '../store'
+
 
 const apiClient = axios.create({
     baseURL: server,
@@ -13,64 +15,60 @@ const apiClient = axios.create({
 
 export default {
 
-    loginUser(data) {
-        return apiClient.post('/auth/login', data)
-    },
-
     addNewUser(data) {
-        return apiClient.post('/users', data)
+        return apiClient.post('/users', data, {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     getUserList() {
-        return apiClient.get('/users')
+        return apiClient.get('/users', {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     editUser(id, data) {
-        return apiClient.put('/users/update/' + id, data)
+        return apiClient.put('/users/update/' + id, data, {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     editDevice(deviceId, data) {
-        return apiClient.put('/devices/update/' + deviceId, data)
+        return apiClient.put('/devices/update/' + deviceId, data, {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     deleteUser(id) {
-        return apiClient.delete('/users/' + id)
+        return apiClient.delete('/users/' + id, {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     addNewDevice(device) {
 
-        return apiClient.post('/devices', device)
+        return apiClient.post('/devices', device, {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     getUserDeviceList(userId) {
-        return apiClient.get('/devices/' + userId)
+        return apiClient.get('/devices/' + userId, {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     getDeviceList() {
-        return apiClient.get('/devices')
+        return apiClient.get('/devices', {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     deleteDevice(deviceId) {
-        return apiClient.delete('/devices/' + deviceId)
+        return apiClient.delete('/devices/' + deviceId, {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     changeDeviceStatus(deviceId, data) {
-        return apiClient.put('/devices/status/' + deviceId, data)
+        return apiClient.put('/devices/status/' + deviceId, data, {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
     getDevicesStats() {
-        return apiClient.get('/devices/statistics')
+        return apiClient.get('/devices/statistics', {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
     getUsersStats() {
-        return apiClient.get('/users/statistics')
+        return apiClient.get('/users/statistics', {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     getRecentlyAddedDevices() {
-        return apiClient.get('/devices/recent')
+        return apiClient.get('/devices/recent', {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     getRecentlyLoggedIn() {
 
-        return apiClient.get('/users/recent')
+        return apiClient.get('/users/recent', {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     },
 
     signInUser(userData) {
@@ -78,6 +76,6 @@ export default {
     },
 
     signOutUser() {
-        apiClient.post('/auth/logout')
+        apiClient.post('/auth/logout', {headers: {Authorization: "Bearer " + state.getters.tokenId}})
     }
 }
